v0.1 indev update 2
Added failsafe, if messaging fails it will attempt it again.
Added icons on tabcontrols
Added try catch on a couple of functions to prevent program crashes.
Working on new modules: Emails, PSEXEC.
User can overwrite a file if they are attempting to name a rmsg file to a file that already exists.
Organised the code a bit.
Prevents forward slash / from being entered into the message box field, this causes an escape and also errors out if done incorrectly.