v0.1 update 8
This is pretty much a bunch of internal upgrades to make each part of the program into their respective classes. Also renamed some files.
Changes:
Created BroadcastHandler class and Enums which is used for each broadcast buffer and RMCManagerForm.
Created the BroadcastController class, this will control all broadcast modules from now on, and will monitor if they are hung. It will create the modules to run them and will coordinate the modules in a asynchronous way.
Separated the Broadcast History files so that PC, email and PSEXEC files are different and are not saved all in one.
Refactored the PCBroadcastModule code so that it will print to the RMC logs about the elapsed time.