Added "panic" feature and refactored code for better clarity
This commit includes several significant changes. A new "panic" feature has been added that allows the program to immediately start broadcasting a predefined emergency message when started with the "PANIC" argument (This however needs to be tested before being approved for v0.1 release). The variable `isScheduledBroadcast` has been replaced with `dontPromptClosureMessage` to better reflect its purpose. The `RMCManager_FormClosing` and `RMCManager_Shown` methods have been updated to handle this new variable. A new method `CloseAfterAllModulesAreFinished` has been added to check if all broadcast modules have finished running. In `BroadcastController.cs`, the private enum `Module` was replaced with `RMCEnums` and the method `AreAnyModulesRunning` was duplicated and renamed to `AreAnyModulesRunningPrivate`. In `RMC IO Manager.cs`, the code was updated to only add the WOL list if it is not empty to fix a parsing error with RMC IO Manager. Two AutoHotkey scripts have been added to activate the "panic" feature. Some resources were removed and added in `Resources.resx` and `Resources.Designer.cs`.