Major Internal Update
[Changes]
IO Manager completely redone, RMC now uses XML formatted RMSG files instead of using a StringBuilder and other functions to get the values of the RMSG parameters. This also fixed the issue of the GetValueFromSection throwing a format exception when getting the value from EmailPort.
Encryption added to RMSG email password. (This encryption is user based DPAPI)
Moved Renew IP button into the log tab instead of on the main tab at the top of the main window.

[TODO]:
(1) FIXME: Why is this function only running the PC module? It should check if the other modules are enabled and run them as well.
Add PSexec saves to the IO manager.