Password Security Update
FIXED: Issue (3) [Type: Security] - 03/08/24 - The EmailPassword is encrypted but it is passed to the IO manager in plain text to be encrypted. This might be a risk since the unencrypted value is passed to different classes. Encrypt the EmailPassword before passing it to the IO manager. (This is now fixed, EncryptionHandler has been removed since that class is only used once, the password gets encrypted by the main form class before giving the encrypted password to the IOManager)
Added OpenLinkOrFolder into IOManager since that code was repeated a bit in the main form.